Refactor SchedulePageModal component
